gdpr compliant clean up, unify alembic
