--- alembic/versions/9a224ef654b3_add_conversation_and_file_lifespan.py
@@ -1,30 +1,0 @@
-"""add conversation and file lifespan
-
-Revision ID: 9a224ef654b3
-Revises: 7aa933ec5de9
-Create Date: 2024-11-25 13:09:16.448641
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '9a224ef654b3'
-down_revision = '7aa933ec5de9'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('organization', sa.Column('conversation_lifespan_days', sa.Integer(), nullable=True))
-    op.add_column('organization', sa.Column('file_lifespan_days', sa.Integer(), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('organization', 'file_lifespan_days')
-    op.drop_column('organization', 'conversation_lifespan_days')
-    # ### end Alembic commands ###
