+++ alembic/versions/eb5ecbee5090_add_conversation_and_file_lifespan.py
@@ -1,0 +1,44 @@
+"""add conversation and file lifespan
+
+Revision ID: eb5ecbee5090
+Revises: 7aa933ec5de9
+Create Date: 2024-11-26 11:03:23.883173
+
+"""
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = "eb5ecbee5090"
+down_revision = "7aa933ec5de9"
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column(
+        "organization",
+        sa.Column("conversation_lifespan_days", sa.Integer(), nullable=True),
+    )
+    op.add_column(
+        "organization",
+        sa.Column(
+            "file_lifespan_days", sa.Integer(), nullable=True, server_default="14"
+        ),
+    )
+    op.drop_column("organization", "gdpr_compliant")
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column(
+        "organization",
+        sa.Column("gdpr_compliant", sa.BOOLEAN(), autoincrement=False, nullable=True),
+    )
+    op.drop_column("organization", "file_lifespan_days")
+    op.drop_column("organization", "conversation_lifespan_days")
+    # ### end Alembic commands ###
